refactor file directory and insert data into db
