+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import "~/styles/globals.css";
 import { GeistSans } from "geist/font/sans";
 import { type Metadata } from "next";
 import { ClerkProvider } from "@clerk/nextjs";
-import { TopNav } from "~/app/_component/topnav";
+import { TopNav } from "~/components/topnav";
 
 export default function RootLayout({
   children,
