+++ src/components/investment_history.tsx
